feat: support recursively rendering nested tables within the TableView component (#1140)
* feat: toggle columns

* feat: expandable rows

* linting

* remove test code

* fix type checking errors

* style: linting

* style: linting
